requirements.txt
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
openai==1.59.8
bs4==0.0.2
faiss-cpu==1.9.0.post1
python-dotenv==1.0.1
tqdm==4.66.4
tiktoken==0.7.0
langchain_chroma==0.1.2
langchainhub==0.1.20